refactor thinking
